Warn for too-high temperatures.
As noted in the code comments for computing the momentum sampling scale
factors in init_species(), the sampling algorithm works down to z = m/T
~ 0.7 for bosons, which corresponds to T ~ 0.192 GeV for the lightest
boson (pi0).  Above this temperature, the sampling envelope partially
falls below the target distribution, so the actual sampled distribution
will be wrong.  However, there is really no good reason to produce
hadrons above this temperature, so it's not much of a limitation.  If
someone mistakenly inputs a temperature in MeV, this warning will notify
them.